policy

+++ app/Http/Controllers/ThreadsController.php
@@ -119,8 +119,17 @@ class ThreadsController extends Controller
      */
     public function destroy($channel,Thread $thread)
     {
+
+        $this->authorize('update',$thread);
+//        if ($thread->user_id != auth()->id()){
+//            abort(403,'you shall not pass');
+//        }
         $thread->replies()->delete();
         $thread->delete();
+
+        if (request()->wantsJson()){
+            return response([],204);
+        }
         return redirect('/threads');
     }
 }
